ai travel planner/requirements.txt
streamlit
requests
python-dotenv
requests